+++ changelog.md
@@ -1,5 +1,7 @@
 # Next
 - [BUG] Fixed issue with `order: sequelize.literal('string')`
+- [FEATURE] add `clone: true` support to `.get()`. Is needed when using `delete` on values from a `.get()` (`toJSON()`, `this.values`). (.get() is just a reference to the values for performance reasons when there's no custom getters or includes)
+- [FEATURE] add `sequelize.escape(value)` convenience method
 
 # 2.0.0-rc6
 - [BUG] Fixed issue with including by association reference and where