+++ changelog.md
@@ -13,6 +13,7 @@
 - [FEATURE] When updating an instance `_previousDataValues` will now be updated after `afterUpdate` hooks have been run rather than before allowing you to use `changed` in `afterUpdate`
 - [BUG] Sequelize will no longer fail on a postgres constraint error not defined by Sequelize
 - [FEATURE] It's now possible to pass an association reference to include. `var Owner = Company.belongsTo(User, {as: 'owner'}; Company.findOne({include: [Owner]});`
+- [BUG] Fixes crash with `findAll({include: [Model], order: sequelize.literal()})`
 
 #### Backwards compatability changes
 - When updating an instance `_previousDataValues` will now be updated after `afterUpdate` hooks have been run rather than before allowing you to use `changed` in `afterUpdate`