+++ changelog.md
@@ -5,6 +5,7 @@
 - [BUG] Fixes crash with `findAll({include: [Model], order: sequelize.literal()})`
 - [FEATURE] Now possible to pass `createdAt` and `updatedAt` values to `Model.create`/`Model.bulkCreate` when using silent: true (when importing datasets with existing timestamps)
 - [FEATURE] `instance.update()` using default fields will now automatically also save values provided via `beforeUpdate` hooks
+- [BUG] Fixed bad SQL when updating a JSON attribute with a different `field`
 
 #### Backwards compatability changes
 - `instance.update()` using default fields will now automatically also save values provided via `beforeUpdate` hooks